services/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
requests==2.31.0
python-dotenv==1.0.1
python-telegram-bot==13.15
aiohttp==3.9.5
